apps/backend/requirements.txt
django
channels
daphne
django-cors-headers
djangorestframework-simplejwt
beautifulsoup4
requests

